+++ bigtable/tests/unit/test_cluster.py
@@ -233,8 +233,8 @@ class TestCluster(unittest.TestCase):
         self.assertEqual(cluster.location, LOCATION)
 
     def test_create(self):
+        from google.api.core import operation
         from google.longrunning import operations_pb2
-        from google.cloud.future import operation
         from google.cloud.bigtable._generated import (
             bigtable_instance_admin_pb2 as messages_v2_pb2)
         from tests.unit._testing import _FakeStub
@@ -275,8 +275,8 @@ class TestCluster(unittest.TestCase):
 
     def test_update(self):
         import datetime
+        from google.api.core import operation
         from google.longrunning import operations_pb2
-        from google.cloud.future import operation
         from google.protobuf.any_pb2 import Any
         from google.cloud._helpers import _datetime_to_pb_timestamp
         from google.cloud.bigtable._generated import (
